requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.4
django-braces==1.14.0
django-embed-video==1.4.0
django-memcache-status==1.3
djangorestframework==3.12.2
idna==2.10
python-memcached==1.59
python3-memcached==1.51
pytz==2020.5
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
